pyproject.toml
[project]
name = "GenerativeKImitPython"
version = "1.1"
description = "Das ist unser Repository für den Kurs 'Generative KI mit Python'"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "chardet>=5.2.0",
    "ipykernel>=6.29.5",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.3",
    "langchain-experimental>=0.3.3",
    "langchain-faiss>=0.1.1",
    "langchain-ollama>=0.3.0",
    "langchain-pymupdf4llm>=0.4.1",
    "langchain-huggingface>=0.1.2",
    "langchain-openai==1.1.10",
    "langchain>=0.3.4",
    "langgraph>=0.2.39",
    "matplotlib>=3.9.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "markdown>=3.7",
    "sentencepiece>=0.2.1",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
    "pytesseract>=0.3.13",
    "pypdf>=6.4.0",
    "youtube-transcript-api>=1.2.3",
    "pytube>=15.0.0",
    "wikipedia>=1.4.0",
    "faiss-cpu>=1.13.0",
    "unstructured>=0.18.21",
    "python-magic>=0.4.27",
    "python-pptx>=1.0.2",
    "google-genai==1.46.0",
    "pydantic>=2.9.0",
    "pydantic>=2.10",
    "langchain-groq>=0.3.1",
    "langchain-tavily>=0.2.17",
    "gradio>=6.6.0",
    "streamlit>=1.54.0",
    "logfire>=4.25.0",
    "seaborn>=0.13.2",
    "langchain-chroma>=1.1.0",
    "sqlmodel>=0.0.37",
    "psycopg2>=2.9.11",
    "langchain-openrouter>=0.2.1",
    "fastmcp>=3.2.4",
]
